+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2025-05-17
+### Added
+- Added List Environments Command: `env:list`.
+- Added dev dependency: `loomlabs/badger`.
+
 ## [1.1.3] - 2025-05-04
 ### Fixed
 - Fixed a typo causing XDebug installation to fail inside the PHP-FPM container.